Refactor code to remove dependency on the prometheus_client library
New versions of the prometheus_client library enforce OpenMetrics naming
which causes metric change in newer versions of Ubuntu.
Due to the way Summary metrics are collected we can no longer use the Summary
object from the library. We also can't easily switch to SummaryMetricFamily
as it's uses metrics registry in a different way.

This change moves away from the prometheus_client library in favour or simple
internal reimplementation.